landing/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.4
django-crispy-forms==1.6.0
django-paypal==0.2.5
django-registration-redux==1.4
haystack==0.36
kombu==3.0.35
Pillow==3.1.1
pytz==2016.2
requests==2.9.1
six==1.10.0
South==1.0.2
winappdbg==1.5
feedparser-5.2.1